requirements.txt
openai-whisper>=20231117
whisper-timestamped>=1.14.0
torch>=2.0.0
torchaudio>=2.0.0

pyannote.audio>=3.1.0
pyannote.core>=5.0.0

librosa>=0.10.0
soundfile>=0.12.0
ffmpeg-python>=0.2.0

pathlib2>=2.3.0  # For older Python versions
typing-extensions>=4.0.0

dotenv